move test to pkce package
