+++ security-oauth2/src/test/groovy/io/micronaut/security/oauth2/endpoint/authorization/request/pkce/OpenIdAuthorizationRedirectWithJustOpenIdPkceSpec.groovy
@@ -1,4 +1,4 @@
-package io.micronaut.security.oauth2.endpoint.authorization.request
+package io.micronaut.security.oauth2.endpoint.authorization.request.pkce
 
 import io.micronaut.core.util.StringUtils
 import io.micronaut.http.HttpHeaders
@@ -6,9 +6,11 @@ import io.micronaut.http.HttpResponse
 import io.micronaut.http.HttpStatus
 import io.micronaut.http.client.DefaultHttpClientConfiguration
 import io.micronaut.http.client.HttpClient
+import io.micronaut.http.client.exceptions.HttpClientResponseException
 import io.micronaut.inject.qualifiers.Qualifiers
 import io.micronaut.security.oauth2.PKCEUtils
 import io.micronaut.security.oauth2.StateUtils
+import io.micronaut.security.oauth2.client.OauthClient
 import io.micronaut.security.oauth2.client.OpenIdClient
 import io.micronaut.security.oauth2.endpoint.authorization.pkce.DefaultPKCEFactory
 import io.micronaut.security.oauth2.routes.OauthController
@@ -19,37 +21,35 @@ import spock.lang.IgnoreIf
 import java.nio.charset.StandardCharsets
 import java.util.regex.Pattern
 
-class OpenIdAuthorizationPKCESpec extends EmbeddedServerSpecification {
-
-    @Override
-    String getSpecName() {
-        'OpenIdAuthorizationPKCESpec'
-    }
+class OpenIdAuthorizationRedirectWithJustOpenIdPkceSpec extends EmbeddedServerSpecification {
 
     @Override
     Map<String, Object> getConfiguration() {
-        Map<String, Object> m = super.configuration + [
+        Map<String, Object> m = super.configuration  + [
                 'micronaut.security.oauth2.pkce.enabled': StringUtils.TRUE,
                 'micronaut.security.authentication': 'cookie',
         ]
         if (System.getProperty(Keycloak.SYS_TESTCONTAINERS) == null || Boolean.valueOf(System.getProperty(Keycloak.SYS_TESTCONTAINERS))) {
             m.putAll([
-                "micronaut.security.oauth2.clients.keycloak.openid.issuer": Keycloak.issuer,
-                "micronaut.security.oauth2.clients.keycloak.client-id": Keycloak.CLIENT_ID,
-                "micronaut.security.oauth2.clients.keycloak.client-secret": Keycloak.clientSecret,
+                    "micronaut.security.oauth2.clients.keycloak.openid.issuer": Keycloak.issuer,
+                    "micronaut.security.oauth2.clients.keycloak.client-id": Keycloak.CLIENT_ID,
+                    "micronaut.security.oauth2.clients.keycloak.client-secret": Keycloak.clientSecret,
             ])
         }
         m
     }
 
     @IgnoreIf({ System.getProperty(Keycloak.SYS_TESTCONTAINERS) != null && !Boolean.valueOf(System.getProperty(Keycloak.SYS_TESTCONTAINERS)) })
-    void "test authorization redirect with openid and pkce"() {
+    void "test authorization redirect with just openid and pkce"() {
         given:
+        Pattern VALID_CODE_CHALLENGE_PATTERN = Pattern.compile('^[0-9a-zA-Z\\-\\.~_]+$')
         HttpClient client = applicationContext.createBean(HttpClient.class, embeddedServer.getURL(), new DefaultHttpClientConfiguration(followRedirects: false))
 
         expect:
         applicationContext.findBean(OpenIdClient, Qualifiers.byName("keycloak")).isPresent()
+        !applicationContext.findBean(OauthClient, Qualifiers.byName("twitter")).isPresent()
         applicationContext.findBean(OauthController, Qualifiers.byName("keycloak")).isPresent()
+        !applicationContext.findBean(OauthController, Qualifiers.byName("twitter")).isPresent()
 
         when:
         HttpResponse response = client.toBlocking().exchange("/oauth/login/keycloak")
@@ -60,15 +60,14 @@ class OpenIdAuthorizationPKCESpec extends EmbeddedServerSpecification {
         location.startsWith(Keycloak.issuer + "/protocol/openid-connect/auth")
         location.contains("scope=openid email profile")
         location.contains("response_type=code")
-        location.contains("client_id=$Keycloak.CLIENT_ID")
         location.contains("redirect_uri=http://localhost:" + embeddedServer.getPort() + "/oauth/callback/keycloak")
-
-        and: 'PKCE cookie is present'
+        location.contains("client_id=$Keycloak.CLIENT_ID")
         response.getCookie("OAUTH2_PKCE").isPresent()
 
         when:
         Map<String, String> queryValues = StateUtils.queryValuesAsMap(location)
         String state = StateUtils.decodeState(queryValues)
+
         then:
         state.contains('"nonce":"')
         state.contains('"redirectUri":"http://localhost:' + embeddedServer.getPort() + '/oauth/callback/keycloak"')
@@ -78,9 +77,16 @@ class OpenIdAuthorizationPKCESpec extends EmbeddedServerSpecification {
         String codeChallengeMethod = PKCEUtils.getCodeChallengeMethod(queryValues)
 
         then:
-        codeChallenge
-        Pattern.compile('^[0-9a-zA-Z\\-\\.~_]+$').matcher(codeChallenge).matches()
+        !codeChallenge.isEmpty()
+        VALID_CODE_CHALLENGE_PATTERN.matcher(codeChallenge)
         codeChallengeMethod == "S256"
         DefaultPKCEFactory.deriveCodeVerifierChallenge(response.getCookie("OAUTH2_PKCE").get().getValue()) == codeChallenge
+
+        when:
+        client.toBlocking().exchange("/oauth/login/twitter")
+
+        then:
+        HttpClientResponseException ex = thrown()
+        ex.response.status.code == 401
     }
 }
